UML-140 : EAnnotation creation/edition/deletion
This commit introduces a new service 'getSemanticElements()" which
leverage a switch to efficiently find the elements to add in the
property view based on the type of the current object.

It also changed pretty much everywhere in the .odesign the references to
Acceleo expressions to compute those elements into a call to this
service.