UML-141 : Support for Default value on Properties
This commit introduces a new menu to create the Default Value of a given
Property.
The default value is also one of the semantic elements so that the end
user an edit it through the property view.
A "Remove" action appears on a property in the "Default Value" submenu
if one is already here.