UML-170: Use Create Instance tool to solve refresh problem
In Object diagram when a Java call is done to create an instance in an
unsynchronized diagram, the new view associated to the new instance
does not appear on the diagram. By using the Create instance tool,
VP will refresh automatically.

In Composite diagram, a same problem occurs when using a select tool
to create a property. Solution was to replace the call of a select tool
by a call to a create node that defines a selection variable wizard.