UML-190 Unsynchronized Use case diagram
"Subject" in a use case diagram can now be any kind of Classifier
(except a UseCase or an Actor themselves) and not only Component like it
was before.
To support that a new section in the palette contains tools to set
Artifact, Class, Collaboration, Component, DataType or Interface as
subjects

drag&dropping any of these elements in the diagram should create the
subject view
drag&dropping a use case or an actor should create the corresponding
shape

the "add element" in the palette have the same behavior.