UML-282: Documentation Table
A new viewpoint is defined which extends the standard UML
representations. This viewpoint is named 'UML Extensions' and
provide a new table representation. This table contains two columns :
- Column 1 contains all the elements defined in a package
- Column 2 contains the comment associated to the UML element.
As a UML element can contain many UML comments, the table allows to
update only the first comment in the ownedComment list.