UML-284 : UseCase to classifier table editor
A new representation is added to create a cross table between all the
classifiers defined in a UML package and all the use cases.
It is possible to edit directly the table and lin/unlink a use case
to a classifier.

The TravelAgency example is edited to show the new Cross table
representation.