UML-184 : Fix copy/paste
A generic tool is implemented to be able to create a view on copy
paste action on unsynchronized diagram.
A "Paste" tool was defined in the odesign, in the reused description
diagram. This tool calls a java service defined in the UIServices.
The service attaches the copied element to the valid parent container
and create the view of the copied element into the target container
which could be a diagram or a container mapping.
Then the tool was reused by all the Existing elements section of
the unsynchronized diagram.