UML-301 : Fix comment creation tool weird behavior
In Acceleo 3 the word 'comment' is a reserved word so the Acceleo
expression of the creation tool is sometime badly interpreted. Just
change the variable name to 'newComment' solve this issue.
This patch also add the possibility to ling a comment to any kind
of elements visible in the Composite Structure diagram.