Package Containment Diagram : tweaks
The Package and Component tools may now be applied directly on a Package
instance, creating a subPackage
The "Containment" tool has been renamed to Move
The background color of the Packages slightly change depending on the
level of depth, getting darker and darker.
The containment relationship is displayed through an edge
A specific filter is here to hide those edges if needed.