UML-322 Add Archetypes layers in Class Diagram
First step introducing a layer to use the UML in color methodology.
The layer provides specific tools to create archetyped classes. These
tools might also be applied on existing classes and will add the marker
to differenciate between archetypes.

Keywords are used to mark the Classes (and not stereotypes as applying a
profile for this looked overkill)