Fix #357 Add Dashboard viewpoint and associated menus
A new viewpoint was added : "UML Dashboard". This viewpoint defines
only one new representation which is the "Dashboard" diagram.
The Dashboard is created automatically at session start on the Model
element and only one dashboard could be created per project.
This diagram contains three main parts:
- Create : List all the available representations and create a new
one thanks to a double click on the representation kind
- Navigate : List all the existing representation and open it thanks
to a double clilck on the representation name
- Help : List the different main sections of the help and double click
to open it (needs to be fixed later when documentation will be updated)

Some new toolbar buttons where also added to the Modeling perspective:
- Create new UML project
- Create new UML model
- Open dashboard : this one shows a sub list off al lthe current
active projects defining a dashboard