Fix #75 Redefine UML viewpoints
Available viewpoints are :
Capture : To capture the actors and their interactions with the application and define the main structure of the application.
Design : To work on the fine grained structure and the behaviours of the application.
Review : To document, get some overview and metrics of the application.
Extend : To extend the UML standard for example by defining profile.