Fix #487 Represent call operation pins according to operation parameters
As the pins must be updated accordingly to the operation parameters,
a resource listener has been implemented to listen the pin/paramter
updates. If the model become inconsistent, decorator are added on
the pin or the call operation action. And new validation checks were
implemented to report call operation action inconsistency in the
problem view.