Fix #551 Existing comments are visible when creating new representation
As now, the layer to display the comments is disabled by default.
So, when you create a new diagram, no comment is shown.
Besides, when you active this layer, only the "useful" comments are
displayed: the ones on the diagram itself and the ones attached to the
current visible graphical elements.

This fix applies on all the diagram kind representations.