Fix#780 Add branch or element, drop from model
Fix add exisiting element:
if an element is selected an is father is not visible, the element is add
to the diagram intead of the element where the tool was applied.
It was not possible to add an element elsewere of his father.

Two services are created:
- to add directly in the selected element
- to add an element from selected branch
If a branch is selected all elements are added with same hierarchy.

Fix drop from model:
when an element is drop from model it is only add to diagram but the
element is no more moved in model.