Fix #775 UML 2.5 display aggregation kind
In uml 2.5 the display of the Aggregation kind is defined as follow and
was taken into account.

abstract UML 2.5 chapter 11.5.4:
"A binary Association may have one end with aggregation =
AggregationKind::shared or aggregation = AggregationKind::composite.
When one end has aggregation = AggregationKind::shared a hollow diamond
is added as a terminal adornment at the end of the Association line
opposite the end marked with aggregation = AggregationKind::shared. The
diamond shall be noticeably smaller than the diamond notation for
Associations. An Association with aggregation =
AggregationKind::composite likewise has a diamond at the corresponding
end, but differs in having the diamond filled in."