Fix: Evol norme UML2.5 Association properties edit "owned" attribut
In UML 2.5 the following evolution was done: chapter 11.5.4:
abstract:
"Ownership of Association ends by an associated Classifier may be
indicated graphically by a small filled circle, which for brevity we
will term a dot. The dot is to be drawn integral to the graphic path of
the line, at the point where it meets the Classifier, inserted between
the end of the line and the side of the node representing the
Classifier. The diameter of the dot shall not exceed half the height of
the aggregation diamond, and shall be larger than the width of the line.
This avoids visual confusion with the filled diamond notation while
ensuring that it can be distinguished from the line. The dot shows that
the model includes a Property of the type represented by the Classifier
touched by the dot. This Property is owned by the Classifier at the
other end. In such a case it is normal to suppress the Property from the
attributes compartment of the owning Classifier."

The edition of the "Onwed" is take into account from EFF properties
view.