Fix #860 Create use case diagram from a use case
In the UML metamodel a use case can contain another use case. But a use
case diagram is based on a package. So if we want to be able to create a
sub use case diagram from a use case we should be able to create a sub
package associated to the use case.

Change-Id: I69396310f227ef3f24e55103b549a8ee4a8110bc
Signed-off-by: Axel Richard <axel.richard@obeo.fr>