Fix #905 Double click tools to open properties dialog
When a double click occurs on a diagram element, the properties edition
dialog opens.
If the user select an element from the model explorer if the properties
view is opened the properties associated to the element are visible.
This patch also fixes the double click tool on an empty diagram which
will import the existing elements.

Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>